Update diegame.py
This code is intended to simulate the roll of a die similar to while playing snakes and ladders. Here I have used the random module in python to manipulate the the die so that the probability of occurence of 1 and 6 is 98/200 while that of the other four numbers are  1/200 each.